+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/handlers/MigrationHandler.kt
@@ -1,7 +1,7 @@
 package com.github.pool_party.pull_party_bot.commands.handlers
 
-import com.elbekD.bot.Bot
-import com.elbekD.bot.types.Message
+import com.elbekd.bot.Bot
+import com.elbekd.bot.types.Message
 import com.github.pool_party.pull_party_bot.commands.EveryMessageInteraction
 import com.github.pool_party.pull_party_bot.database.dao.ChatDao
 import mu.KotlinLogging
@@ -13,7 +13,7 @@ class MigrationHandler(private val chatDao: ChatDao) : EveryMessageInteraction {
     override suspend fun onMessage(bot: Bot, message: Message) {
 
         val fromChatId = message.chat.id
-        val toChatId = message.migrate_to_chat_id ?: return
+        val toChatId = message.migrateToChatId ?: return
 
         logger.info { "Migration: $fromChatId -> $toChatId" }