+++ src/test/kotlin/com/github/pool_party/pull_party_bot/commands/handlers/AbstractBotTest.kt
@@ -1,16 +1,13 @@
 package com.github.pool_party.pull_party_bot.commands.handlers
 
-import com.elbekD.bot.Bot
-import com.elbekD.bot.types.CallbackQuery
-import com.elbekD.bot.types.Chat
-import com.elbekD.bot.types.ChatMember
-import com.elbekD.bot.types.Message
-import com.elbekD.bot.types.User
-import com.github.pool_party.pull_party_bot.commands.CallbackDispatcher
-import com.github.pool_party.pull_party_bot.commands.EveryMessageProcessor
-import com.github.pool_party.pull_party_bot.commands.handlers.callback.DeleteNodeSuggestionCallback
-import com.github.pool_party.pull_party_bot.commands.handlers.callback.DeleteSuggestionCallback
-import com.github.pool_party.pull_party_bot.commands.handlers.callback.PingCallback
+import com.elbekd.bot.Bot
+import com.elbekd.bot.types.CallbackQuery
+import com.elbekd.bot.types.Chat
+import com.elbekd.bot.types.ChatMember
+import com.elbekd.bot.types.InlineKeyboardMarkup
+import com.elbekd.bot.types.Message
+import com.elbekd.bot.types.User
+import com.github.pool_party.pull_party_bot.commands.initHandlers
 import com.github.pool_party.pull_party_bot.database.AliasCache
 import com.github.pool_party.pull_party_bot.database.Aliases
 import com.github.pool_party.pull_party_bot.database.ChatCache
@@ -18,14 +15,11 @@ import com.github.pool_party.pull_party_bot.database.Chats
 import com.github.pool_party.pull_party_bot.database.Parties
 import com.github.pool_party.pull_party_bot.database.PartyAliasesCache
 import com.github.pool_party.pull_party_bot.database.PartyCache
-import com.github.pool_party.pull_party_bot.database.dao.ChatDaoImpl
-import com.github.pool_party.pull_party_bot.database.dao.PartyDaoImpl
 import io.mockk.MockKAnnotations
+import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import org.flywaydb.core.Flyway
 import org.jetbrains.exposed.sql.Database
@@ -33,48 +27,27 @@ import org.jetbrains.exposed.sql.deleteAll
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.testcontainers.containers.PostgreSQLContainer
 import org.testcontainers.junit.jupiter.Container
-import java.util.concurrent.CompletableFuture
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
+import kotlin.test.assertNotNull
 
 internal abstract class AbstractBotTest {
 
     @MockK
     protected lateinit var bot: Bot
 
-    protected val chat =
+    private val chat =
         Chat(1, "group", null, null, null, null, null, null, null, null, null, null, null, null, null, null, null)
 
-    protected val user = User(-1, false, "first name", null, "admin", null, null, null, null)
+    private val user = User(-1, false, "first name", null, "admin", null, null, null, null)
 
-    protected val chatMember = ChatMember(
-        user,
-        "status",
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null
-    )
+    private val chatMember = ChatMember.Owner("creator", user)
 
     protected val message = Message(
         1,
         user,
         null,
-        System.currentTimeMillis().div(1000).toInt(),
+        System.currentTimeMillis().div(1000).toLong(),
         chat,
         null,
         null,
@@ -90,27 +63,27 @@ internal abstract class AbstractBotTest {
         null,
         null,
         null,
+        listOf(),
+        null,
+        null,
+        null,
+        listOf(),
+        null,
+        null,
+        null,
+        null,
+        null,
+        listOf(),
         null,
         null,
         null,
         null,
         null,
         null,
+        listOf(),
         null,
         null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
+        listOf(),
         null,
         null,
         null,
@@ -124,46 +97,19 @@ internal abstract class AbstractBotTest {
         null
     )
 
-    private val chatDao = ChatDaoImpl()
-
-    private val partyDao = PartyDaoImpl()
-
-    private val everyMessageProcessor =
-        EveryMessageProcessor(listOf(MigrationHandler(chatDao), ImplicitPartyHandler(partyDao)))
-
     private lateinit var everyMessageAction: suspend (Message) -> Unit
 
     private lateinit var callbackAction: suspend (CallbackQuery) -> Unit
 
-    private val commandActions = mutableMapOf<String, suspend (Message, String?) -> Unit>()
+    private val commandActions = mutableMapOf<String, suspend (Pair<Message, String?>) -> Unit>()
 
-    private val commands = listOf(
-        StartCommand(),
-        ListCommand(partyDao, chatDao),
-        PartyCommand(partyDao),
-        DeleteCommand(partyDao, chatDao),
-        ClearCommand(chatDao),
-        CreateCommand(partyDao, chatDao),
-        AliasCommand(partyDao, chatDao),
-        ChangeCommand(partyDao, chatDao),
-        AddCommand(partyDao, chatDao),
-        RemoveCommand(partyDao, chatDao),
-        RudeCommand(chatDao),
-        FeedbackCommand(),
-    )
-
-    private val callbacks = listOf(
-        DeleteNodeSuggestionCallback(partyDao),
-        DeleteSuggestionCallback(partyDao),
-        PingCallback(partyDao),
-    )
-
-    private val callbackDispatcher = CallbackDispatcher(callbacks.associateBy { it.callbackAction })
+    protected var lastCallbackData: String? = null
 
     @BeforeTest
     fun setupMock() {
         MockKAnnotations.init(this, relaxed = true, relaxUnitFun = true)
 
+        // start and migrate database
         container.start()
         Database.connect(
             "${container.jdbcUrl}&gssEncMode=disable",
@@ -172,19 +118,32 @@ internal abstract class AbstractBotTest {
         )
         Flyway.configure().dataSource(container.jdbcUrl, container.username, container.password).load().migrate()
 
-        every { bot.sendMessage(allAny(), allAny()) } answers {
+        // mock basic functions
+        coEvery { bot.sendMessage(allAny(), allAny()) } answers {
             println(">> ${secondArg<String>()}")
-            CompletableFuture.completedFuture(message)
+            message
         }
-        every { bot.getChatAdministrators(chat.id) } returns CompletableFuture.completedFuture(arrayListOf(chatMember))
-        every { bot.onCommand(any(), any()) } answers { commandActions[firstArg()] = secondArg() }
+        coEvery { bot.getChatAdministrators(any()) } returns arrayListOf(chatMember)
+        coEvery { bot.onCommand(any(), any()) } answers { commandActions[firstArg()] = secondArg() }
         every { bot.onMessage(any()) } answers { everyMessageAction = firstArg() }
         every { bot.onCallbackQuery(any()) } answers { callbackAction = firstArg() }
-        every { bot.deleteMessage(any(), any()) } returns CompletableFuture.completedFuture(true)
+        coEvery { bot.deleteMessage(any(), any()) } returns true
 
-        commands.forEach { it.onMessage(bot) }
-        everyMessageProcessor.onMessage(bot)
-        callbackDispatcher.onMessage(bot)
+        // extract functions
+        runBlocking { bot.initHandlers() }
+
+        // mock callbacks
+        coEvery { bot.sendMessage(allAny(), allAny()) } answers {
+            val inlineKeyboardMarkup = arg(9) as? InlineKeyboardMarkup
+            println(">> ${secondArg<String>()}\n*button*: $inlineKeyboardMarkup")
+            lastCallbackData = inlineKeyboardMarkup?.inlineKeyboard?.singleOrNull()?.singleOrNull()?.callbackData
+            message
+        }
+        coEvery { bot.answerCallbackQuery(allAny(), allAny()) } answers {
+            println("^^ ${secondArg<String>()}")
+            true
+        }
+        coEvery { bot.answerCallbackQuery(allAny()) } returns true
     }
 
     @AfterTest
@@ -203,47 +162,49 @@ internal abstract class AbstractBotTest {
 
     // bot interaction test DSL
 
-    protected fun callback(callbackData: String) {
-        runBlocking {
-            GlobalScope.launch {
-                callbackAction(CallbackQuery("id", user, message, null, "chat instance", callbackData, null))
-            }.join()
-        }
+    protected fun callback(callbackData: String) = runBlocking {
+        callbackAction(CallbackQuery("id", user, message, null, "chat instance", callbackData, null))
     }
 
-    protected operator fun String.unaryMinus() {
-        println("<< $this")
+    protected fun clickButton() = runBlocking {
+        assertNotNull(lastCallbackData)
+        callbackAction(CallbackQuery("id", user, message, null, "chat instance", lastCallbackData!!, null))
+    }
+
+    protected operator fun String.unaryMinus() = runBlocking {
+        println("<< ${this@unaryMinus}")
         val split = split(" ")
-        runBlocking {
-            GlobalScope.launch {
-                val command = commandActions[split[0]]
-                val currentMessage = message.copy(text = this@unaryMinus)
+        val command = commandActions[split[0]]
+        val currentMessage = message.copy(text = this@unaryMinus)
 
-                if (command != null) {
-                    command(currentMessage, split.drop(1).joinToString(" "))
-                } else {
-                    everyMessageAction(currentMessage)
-                }
-            }.join()
+        if (command != null) {
+            command(currentMessage to split.drop(1).joinToString(" "))
+        } else {
+            everyMessageAction(currentMessage)
         }
     }
 
-    protected fun verifyMessages(chatId: Long, text: String, exactly: Int = 1, replyTo: Long? = null) {
+    protected fun verifyMessage(exactly: Int = 1, replyTo: Long? = null, matcher: (String) -> Boolean) {
         coVerify(exactly = exactly) {
-            bot.sendMessage(chatId, text, any(), any(), any(), any(), replyTo ?: any(), any(), any())
+            bot.sendMessage(
+                match { it.toString() == chat.id.toString() },
+                match(matcher),
+                any(),
+                any(),
+                any(),
+                any(),
+                any(),
+                replyTo ?: any(),
+                any(),
+                any(),
+            )
         }
     }
 
-    protected fun verifyMessage(matcher: (String) -> Boolean) {
-        coVerify {
-            bot.sendMessage(chat.id, match(matcher), any(), any(), any(), any(), any(), any(), any())
-        }
-    }
+    protected fun verifyContains(vararg substrings: String, exactly: Int = 1) =
+        verifyMessage(exactly = exactly) { message -> substrings.all { it in message } }
 
-    protected fun verifyContains(vararg substrings: String) =
-        verifyMessage { message -> substrings.all { it in message } }
-
-    protected operator fun String.unaryPlus() = verifyMessages(chat.id, this)
+    protected operator fun String.unaryPlus() = verifyMessage { it == this }
 
     companion object {
         @Container